Tue, Aug 2, 2022, 3:28 PM +00:00

+++ scrabble/scrabble.c
@@ -8,7 +8,7 @@
 int POINTS[] = {1, 3, 3, 2, 1, 4, 2, 4, 1, 8, 5, 1, 3, 1, 1, 3, 10, 1, 1, 1, 1, 4, 4, 8, 4, 10};
 
 int compute_score(string word);
-int s = 0;
+float s = 0;
 int main(void)
 {
 
@@ -21,7 +21,7 @@ int main(void)
     //int score2 = compute_score(word2);
 
     // TODO: Print the winner
-    printf ("Score = %i", s);
+    printf ("Score = %f", s);
     //if (int score1 >> int score2)
     //printf ("Player 1 wins!\n");
     //if (int score1 << int score2)
